Service functional to specifications
The robot continues to rotate after the `__rotate` function has exited,
to an angle proportional to the angular velocity of the rotation. This
extra angle has not been accounted for but the last commit before the
tag `odom-topic` reports the current yaw at the beginning and end of
each service call. The current difference is about 3 degrees in the
direction of rotation.